finished orders management

+++ src/main/java/com/siewe/inventorymanagementsystem/repository/OrderedProduitRepository.java
@@ -5,9 +5,9 @@ import com.siewe.inventorymanagementsystem.model.Orders;
 import com.siewe.inventorymanagementsystem.model.Produit;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderedProduitRepository extends JpaRepository<OrderedProduit, Long> {
+public interface OrderedProduitRepository extends JpaRepository<OrderedProduit, Integer> {
 
-    default OrderedProduit findOne(Long id) {
+    default OrderedProduit findOne(Integer id) {
         return (OrderedProduit) findById(id).orElse(null);
     }
     OrderedProduit findByOrdersAndProduct(Orders orders, Produit produit);
